Adding additional database driver functionality. Folding container logic back into modified structure.

+++ NEsper/NEsper.Common/common/client/configuration/runtime/ConfigurationRuntimeParser.cs
@@ -353,7 +353,7 @@ namespace com.espertech.esper.common.client.configuration.runtime
 
                         if (subElement.Attributes.GetNamedItem("timeout-msec") != null) {
                             var timeoutMSecText = subElement.Attributes.GetNamedItem("timeout-msec").InnerText;
-                            var timeoutMSec = long.Parse(timeoutMSecText);
+                            var timeoutMSec = int.Parse(timeoutMSecText);
                             runtime.Threading.NamedWindowConsumerDispatchTimeout = timeoutMSec;
                         }
